backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0,<3.15
django-extensions>=3.1.3
drf-spectacular
Pillow
django_extensions
psycopg2-binary>=2.9.3,<3.0
django-crispy-forms
crispy-bootstrap4
django-cors-headers